Split router routes list to files
